Done w/ UI, need to finish dev data

+++ db/migrate/20220203225215_create_logs.rb
@@ -5,7 +5,7 @@ class CreateLogs < ActiveRecord::Migration[6.0]
       t.integer :user_id
       t.datetime :started_at
       t.integer :duration
-      t.integer :distance
+      t.decimal :distance
       t.integer :calories
 
       t.timestamps
